refactor: 주행 리포트 DTO 분리 및 리뷰 반영
