+++ analysis-service/src/main/kotlin/com/parkit/analysis/driving/service/SensorLogService.kt
@@ -4,6 +4,7 @@ import com.parkit.analysis.driving.persistence.document.SensorLogDocument
 import com.parkit.analysis.driving.persistence.repository.SensorLogMongoRepository
 import com.parkit.analysis.kafka.dto.ParkingSensorDto
 import org.springframework.stereotype.Service
+import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import java.time.Clock
 import java.time.Instant
@@ -33,4 +34,7 @@ class SensorLogService(
 		)
 		return sensorLogRepository.save(doc)
 	}
+
+	fun getSensorLogs(sessionId: String, limit: Long): Flux<SensorLogDocument> =
+		sensorLogRepository.findBySessionIdOrderByTimeAsc(sessionId).take(limit)
 }
