+++ analysis-service/src/main/resources/application.yml
@@ -10,7 +10,7 @@ spring:
       host: localhost
       port: 6379
     mongodb:
-      uri: mongodb://mongodb-service:27017/dbname
+      uri: ${PARKIT_MONGODB_URI:mongodb://mongodb-service:27017/analysis_db}
   kafka:
     bootstrap-servers: my-kafka.parkit.svc.cluster.local:9092
     properties:
